Fix: 종료 큐 원소 크기 및 연결 리스트 삭제 오류 수정

+++ IA32Mode/src/Process.c
@@ -64,7 +64,7 @@ void initScheduler(void) {
 	allocProcessTable[pidCountIdx++] = 1;
 	initList(&(scheduler.processList));
 	scheduler.runningProcess = pcb;
-	initQueue(&exitProcessQueue, queueBuffer, EXIT_QUEUE_COUNT, sizeof(QWORD));
+	initQueue(&exitProcessQueue, queueBuffer, EXIT_QUEUE_COUNT, sizeof(int)); // link.id 크기와 일치
 	createProcess((QWORD)garbegeProcessCollector);
 }
